Cleanup: macros & imports (#254)
* move assert_approx_eq macro to test_utils crate

* remove macro_use and limit local macro visibility

* move test-only utils
