+++ dev-tool/Cargo.toml
@@ -27,4 +27,4 @@ indicatif = "0.16.2"
 rayon = "1.5.1"
 
 [dev-dependencies]
-float-cmp = "0.9.0"
+test-utils = { path = "../test-utils" }
