+++ test-utils/Cargo.toml
@@ -8,5 +8,8 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+float-cmp = "0.9.0"
+# to be kept in sync with rubert
+ndarray = "=0.15.3"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
