+++ xayn-ai/Cargo.toml
@@ -30,14 +30,13 @@ rayon = { version = "1.5.1", optional = true}
 
 [dev-dependencies]
 csv = "1.1.6"
-test-utils = { path = "../test-utils" }
-float-cmp = "0.9.0"
 maplit = "1.0.2"
 mockall = "0.10.2"
 once_cell = "1.8.0"
 paste = "1.0.5"
 rstest = "0.11.0"
 rstest_reuse = "0.1.3"
+test-utils = { path = "../test-utils" }
 
 [features]
 multithreaded = ["rayon"]
