+++ xayn-ai/src/lib.rs
@@ -1,6 +1,3 @@
-#[macro_use]
-mod utils;
-
 mod analytics;
 mod coi;
 mod context;
@@ -9,6 +6,7 @@ mod embedding;
 mod error;
 mod ltr;
 mod reranker;
+mod utils;
 
 pub use crate::{
     analytics::Analytics,
@@ -42,10 +40,6 @@ mod tests;
 #[allow(unused_imports)]
 #[rustfmt::skip]
 pub(crate) use rstest_reuse as rstest_reuse;
-
-// Reexport for assert_approx_eq for usage in FFI crates.
-#[doc(hidden)]
-pub use self::utils::ApproxAssertIterHelper;
 
 // Reexport for the dev-tool
 #[doc(hidden)]
