+++ xayn-ai/src/ltr/list_net/ndlayers/activation.rs
@@ -116,6 +116,7 @@ mod tests {
     use ndarray::{arr3, Axis};
 
     use super::*;
+    use test_utils::assert_approx_eq;
 
     #[test]
     fn test_relu_activation_function_works() {
